WORKING ONLY: new create_pp_channel function
This is a working version of this function. It will need to be moved to be called from the Advanced tab. As it's written now, it relies on EE tags being in a template, but that's messy. the new channel and channel fields it is creating need to be able to be built or deleted from the advanced tab area.

Signed-off-by: John Morton <john@johnfmorton.com>